+++ lib/rordan_gramsay/lint/rubocop.rb
@@ -1,35 +1,37 @@
-require_relative '../../rordan_gramsay'
-RordanGramsay.silence_warnings do
-  require 'foodcritic'
-end
+require 'json'
 
 require_relative 'base'
-require_relative '../foodcritic/file_list'
-require_relative '../foodcritic/rule'
+require_relative '../rubocop/file_list'
+require_relative '../rubocop/rule'
 
 module RordanGramsay
   module Lint
     # :nodoc:
-    class Foodcritic < Base
-      attr_reader :results, :file_list, :opts
+    class Rubocop < Base
+      attr_reader :results, :file_list
 
       def initialize(opts = {})
         @opts = default_opts.merge(opts)
-        @file_list = RordanGramsay::Foodcritic::FileList.new
+        @file_list = RordanGramsay::Rubocop::FileList.new
         @linter = nil
 
-        RordanGramsay.silence_warnings do
-          @linter = ::FoodCritic::Linter.new
-          results = @linter.check(@opts.dup)
+        json = JSON.parse(`rubocop --require cookstyle --format json`)
 
-          paths = @linter.send(:specified_paths!, @opts.dup)
-          @linter.send(:files_to_process, paths).each do |f|
-            @file_list << f[:filename]
-          end
+        json['files'].each do |file_json|
+          file = @file_list[::File.expand_path(file_json['path'])]
 
-          results.warnings.each do |warn|
-            rules = @file_list[warn.match[:filename]].rules
-            rules[warn.match[:line].to_i] << RordanGramsay::Foodcritic::Rule.new(warn.rule, warn.failed?)
+          file_json['offenses'].each do |offense_json|
+            offense = Struct.new(:code, :name).new(offense_json['cop_name'], offense_json['message'])
+            # Don't stick with syntax error specifics here. We're supposed to log general rule descriptions
+            offense.name = 'Ruby syntax error' if offense.code == 'Syntax'
+            has_failed = fail_severity?(offense_json['severity'])
+            line_number = begin
+                            offense_json.fetch('location').fetch('line')
+                          rescue KeyError
+                            1
+                          end
+
+            file.rules[line_number] << RordanGramsay::Rubocop::Rule.new(offense, has_failed)
           end
         end
       end
@@ -38,7 +40,7 @@ module RordanGramsay
         @file_list.filenames
       end
 
-      # Outputs the results of Foodcritic linting by filename
+      # Outputs the results of RuboCop linting by filename
       #
       #   <space x2><filename> : <result>
       #   <space x4><line num>: <rule codes>
@@ -72,7 +74,7 @@ module RordanGramsay
       end
 
       def rules
-        rules = RordanGramsay::Foodcritic::RuleList.new
+        rules = RordanGramsay::Rubocop::RuleList.new
         @file_list.each do |file|
           file.rules.each do |(_, ruleset)|
             ruleset.each do |rule|
@@ -88,11 +90,23 @@ module RordanGramsay
 
       def default_opts
         {
-          cookbook_paths: [Dir.pwd],
-          exclude_paths: %w(test/**/* spec/**/* features/**/*),
-          fail_tags: ['any'],
-          tags: []
+          fail_severity: 'error'
         }
+      end
+
+      def fail_severity?(severity)
+        severity_map = {
+          fatal: 0,
+          error: 1,
+          warning: 2,
+          convention: 3,
+          refactor: 4
+        }
+
+        minimum_lvl = severity_map.fetch(@opts['fail_severity']) { 99 }
+        severity_lvl = severity_map.fetch(severity) { 99 }
+
+        minimum_lvl <= severity_lvl
       end
     end
   end