+++ CHANGELOG.md
@@ -5,7 +5,7 @@ means we will never make a backwards-incompatible change within a major version 
 
 ## _[UNRELEASED]_
 
-No changes yet...
+- [BUG] Rubocop linting task didn't actually run Rubocop
 
 ## v1.3.0 (2017-06-08)